requirements.txt
langchain
langchain-community
pandas
openai
langgraph
fastapi
uvicorn
python-dotenv
dotenv
python-multipart
Jinja2
langchain-core
langchain-openai
langchain_chroma
langchain-chroma
pymongo
itsdangerous
jinja2
mysql-connector-python
pymongo[srv]==3.11
gunicorn
twilio
pysqlite3-binary